+++ tools/package-artifacts/src/index.ts
@@ -1,9 +1,10 @@
 export * from './build'
-export * from './catalog'
-export * from './exports'
-export * from './root'
-export * from './registry-output'
-export * from './npm-output'
-export * from './manifest'
+export * from './workspace/catalog'
+export * from './manifest/exports'
+export * from './workspace/root'
+export * from './npm/registry'
+export { parseJSONObject } from './json'
+export * from './npm/pack'
+export * from './manifest/read'
 export * from './process'
-export * from './types'
+export * from './manifest/types'
