+++ tools/package-artifacts/src/npm/pack.ts
@@ -1,6 +1,6 @@
 import * as v from 'valibot'
 
-const jsonObjectSchema = v.record(v.string(), v.unknown())
+import { parseJSON } from '../json'
 
 function isPackageRelativePath(path: string): boolean {
   return (
@@ -28,23 +28,6 @@ const packResultSchema = v.pipe(
 )
 
 export type NpmPackResult = v.InferOutput<typeof packResultSchema>
-
-function parseJSON(text: string, context: string): unknown {
-  try {
-    return JSON.parse(text)
-  } catch {
-    throw new Error(`${context}: invalid JSON`)
-  }
-}
-
-export function parseJSONObject(
-  text: string,
-  context: string
-): v.InferOutput<typeof jsonObjectSchema> {
-  const parsed = v.safeParse(jsonObjectSchema, parseJSON(text, context))
-  if (!parsed.success) throw new Error(`${context}: expected an object`)
-  return parsed.output
-}
 
 export function parseNpmPack(text: string): NpmPackResult {
   const parsed = v.safeParse(packResultSchema, parseJSON(text, 'npm pack'))
