+++ tools/package-artifacts/src/npm/registry.ts
