--- tools/release-packages/src/tarballs.ts
@@ -1,1 +1,0 @@
-export * from '@open-pencil/package-artifacts/tarball'
